Changes to Auth configurations (#68)
I realized while testing things a bit more locally that just having one endpoint was insufficient - flyte-cli and pyflyte currently leverage the gRPC Python Flyte Admin client.  However, discovery needs the HTTP endpoint.  Even though in production both are served on the same endpoint and traffic is handled through ingress, this is not always the case, and we want to support the use-case where gRPC Admin and HTTP Admin are hosted separately.

We were debating just adding a separate config for discovery alone, but ultimately it makes more sense to split out the existing URL config into two objects, especially if the above pattern is something we want to continue to support.  However, in the interest of not making people migrate, and not writing migration code in this PR, we're just introducing the HTTP one for now.

Also, we should lowercase the metadata header before sending it to Admin in the gRPC calls.