Add support for string-format-like sytax for shell task (#792)
* POC: Add support for f-string like sytax for shell task

This commit is a proof of concept adding f-string like syntax for
shell_tasks. This supports using nested types for script inputs, such
as data classes. This change was motivated by the desire to combine
shell_tasks that have multiple inputs with map_tasks which only support
tasks with a single input.

This commit is only a starting point, since it makes some changes to the
shell_task API (adds a template_style field), and modifies some of the default
behavior for ease of implementation (e.g. throwing an error when there
are unused input arguments).

Signed-off-by: Zach Palchick <palchicz@zymergen.com>

* Drop support for old/regex style for doing string interpolation

Signed-off-by: Zach Palchick <palchicz@zymergen.com>